+++ Webstore/src/main/java/org/mentor/dao/ClientDaoImpl.java
@@ -1,7 +1,6 @@
 package org.mentor.dao;
 
 import org.mentor.model.Client;
-import org.mentor.jpa.JPAEntityManager;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
