+++ Webstore/src/main/java/org/mentor/dao/ProductDaoImpl.java
@@ -1,7 +1,6 @@
 package org.mentor.dao;
 
 import org.mentor.model.Product;
-import org.mentor.jpa.JPAEntityManager;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
