+++ Webstore/src/main/java/org/mentor/model/Receipt.java
@@ -1,10 +1,12 @@
 package org.mentor.model;
 
 import lombok.Data;
+import org.springframework.stereotype.Repository;
 
 import javax.persistence.*;
 
 @Data
+@Repository
 @Entity
 public class Receipt {
     @Id
@@ -15,14 +17,14 @@ public class Receipt {
     private Integer bought;
     @Column(name = "total_price")
     private Double totalPrice;
-
+    //Bi-direct(Owner side)
     @ManyToOne(cascade = {
             CascadeType.DETACH, CascadeType.MERGE,
             CascadeType.PERSIST, CascadeType.REFRESH
     })
     @JoinColumn(name = "client_id")
     private Client client;
-
+    //Bi-direct(Owner side)
     @ManyToOne(cascade = {
             CascadeType.DETACH, CascadeType.MERGE,
             CascadeType.PERSIST, CascadeType.REFRESH
@@ -35,5 +37,13 @@ public class Receipt {
     public Receipt(Integer bought, Double totalPrice) {
         this.bought = bought;
         this.totalPrice = totalPrice;
+    }
+
+    public Receipt(Integer id, Integer bought, Double totalPrice, Client client, Product product) {
+        this.id = id;
+        this.bought = bought;
+        this.totalPrice = totalPrice;
+        this.client = client;
+        this.product = product;
     }
 }
